+++ CHANGELOG.md
@@ -4,8 +4,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Silver
 ### Changed
-- World chaos will now balance at a certain level, instead of zero. The level fluctuates over time. 
+- World chaos will now balance at a certain level, instead of zero. The level fluctuates over time.
+- Wild fluffy puff plants will only drop seeds now 
 ### Fixed
 - Crash when Silent Gear not installed [#365]
 - Server crash